cephadm: move byte string manipulation funcs to data_utils.py
Pair-programmed-with: Adam King <adking@redhat.com>
